Revert "bootstrapper - full prerequesites ui"
This reverts commit 9fed9c3e88ed07d4a9e27f20ff15fcc06b1ede67.
