+++ TeslaCam/App.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:TeslaCam"
              xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
-             StartupUri="BootstrapWindow.xaml"
+             StartupUri="MainWindow.xaml"
              ThemeMode="System">
     <Application.Resources>
         <ResourceDictionary>
