--- TeslaCam/BootstrapWindow.xaml
@@ -1,62 +1,0 @@
-﻿<Window x:Class="TeslaCam.BootstrapWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TeslaCam"
-        mc:Ignorable="d"
-        d:DataContext="{d:DesignInstance Type=local:BootstrapWindow}"
-        Title="TeslaCam Player"
-        Height="450"
-        Width="800"
-        Visibility="Hidden"
-        Loaded="Window_Loaded"
-        ContentRendered="Window_ContentRendered">
-    <StackPanel>
-        <!-- Header -->
-        <TextBlock Text="TeslaCam" 
-                   FontSize="36" 
-                   FontWeight="Bold" 
-                   HorizontalAlignment="Center" 
-                   VerticalAlignment="Top" 
-                   Margin="0,20,0,0" />
-
-        <!-- State Content -->
-        <ContentControl Content="{Binding State}">
-            <ContentControl.Resources>
-                <!-- Initializing State -->
-                <DataTemplate DataType="{x:Type local:BootstrapState}">
-                    <TextBlock Text="" />
-                </DataTemplate>
-
-                <!-- Missing Dependencies State -->
-                <DataTemplate DataType="{x:Type local:BootstrapStateMissingDependencies}">
-                    <StackPanel HorizontalAlignment="Center">
-                        <TextBlock Text="Missing Dependencies. Do you want to install them now?" 
-                                   TextAlignment="Center" 
-                                   Margin="0,0,0,10" />
-                        <Button Content="Yes" Command="{Binding InstallDependenciesCommand}" Margin="5" />
-                        <Button Content="No" Command="{Binding CloseCommand}" Margin="5" />
-                    </StackPanel>
-                </DataTemplate>
-
-                <!-- Installing State -->
-                <DataTemplate DataType="{x:Type local:BootstrapStateInstalling}">
-                    <StackPanel HorizontalAlignment="Center">
-                        <TextBlock Text="Installing" TextAlignment="Center" Margin="0,0,0,10" />
-                        <ProgressBar Width="200" Height="20" IsIndeterminate="True" Margin="0,10,0,10" />
-                        <Button Content="Cancel" Command="{Binding CancelCommand}" Margin="5" />
-                    </StackPanel>
-                </DataTemplate>
-
-                <!-- Error State -->
-                <DataTemplate DataType="{x:Type local:BootstrapStateError}">
-                    <StackPanel HorizontalAlignment="Center">
-                        <TextBlock Text="Something went wrong." TextAlignment="Center" Margin="0,0,0,10" />
-                        <Button Content="Close" Command="{Binding CloseCommand}" Margin="5" />
-                    </StackPanel>
-                </DataTemplate>
-            </ContentControl.Resources>
-        </ContentControl>
-    </StackPanel>
-</Window>
